fix: Fix routes for admin rights

+++ ressources/models/ModelResource.py
@@ -55,40 +55,43 @@ class ModelResource(Resource):
         needEmail = request.json.get('needEmail', False)
         if parameters is None:
             parameters = modelDB.parameters
-        model = ModelCVRPApi()
-        model.readFromData(modelDB.data_parameters)
-        model.initModel(parameters['vehicle_max_capacity'])
-        model.modelGurobi.update()
+        if (user != None and (user.username == current_username or user.isadmin)):
+            model = ModelCVRPApi()
+            model.readFromData(modelDB.data_parameters)
+            model.initModel(parameters['vehicle_max_capacity'])
+            model.modelGurobi.update()
 
-        #Create file with solution text
-        with open(str(modelDB.id) + '.sol', 'w') as f:
-            f.write(modelDB.file_solution_text)
+            #Create file with solution text
+            with open(str(modelDB.id) + '.sol', 'w') as f:
+                f.write(modelDB.file_solution_text)
 
-        model.modelGurobi.read(str(modelDB.id) + '.sol')
-        model.optimizeModel(parameters)
-        model.modelGurobi.write(str(modelDB.id) + '.sol')
-        with open(str(modelDB.id) + '.sol') as f:
-            modelDB.file_solution_text = f.read()
-            f.close()
-        os.remove(str(modelDB.id) + '.sol')
-        modelDB.solution = model.getRoutesFromSolution()
-        modelDB.parameters = parameters
-        modelDB.name = name
-        modelDB.last_edit = datetime.datetime.now()
-        db.session.commit()
-        print(model.getRoutesFromSolution())
-        if(needEmail):
-            mailTemplate = MailTemplate(user.email, f"CVRP - Model Execution {modelDB.name} ended", "Execution ended!")
-            mailTemplate.sendMail()
-        return jsonify(model=model_schema.dump(modelDB))
+            model.modelGurobi.read(str(modelDB.id) + '.sol')
+            model.optimizeModel(parameters)
+            model.modelGurobi.write(str(modelDB.id) + '.sol')
+            with open(str(modelDB.id) + '.sol') as f:
+                modelDB.file_solution_text = f.read()
+                f.close()
+            os.remove(str(modelDB.id) + '.sol')
+            modelDB.solution = model.getRoutesFromSolution()
+            modelDB.parameters = parameters
+            modelDB.name = name
+            modelDB.last_edit = datetime.datetime.now()
+            db.session.commit()
+            print(model.getRoutesFromSolution())
+            if(needEmail):
+                mailTemplate = MailTemplate(user.email, f"CVRP - Model Execution {modelDB.name} ended", "Execution ended!")
+                mailTemplate.sendMail()
+            return jsonify(model=model_schema.dump(modelDB))
+        raise NotAllowedError("User not allowed to edit it")
 
     @jwt_required()
     def get(self, modelId):
         current_username = get_jwt_identity() # Get username from current user
         model = Model.query.filter_by(id=modelId).first() # Get model
+        user = Users.query.filter_by(username=current_username).first()
         if model == None:
             raise NotFoundError("Model not found") 
-        if model.user.username == current_username:
+        if user != None and (model.user.username == current_username or user.isadmin):
             return jsonify(model=model_schema.dump(model))
         raise NotAllowedError("User not allowed to access it")
     
@@ -99,7 +102,7 @@ class ModelResource(Resource):
         model = Model.query.filter_by(id=modelId).first() # Get model
         if model == None:
             raise NotFoundError("Model not found") 
-        if user.isadmin or model.user.username == current_username:
+        if user != None and (user.isadmin or model.user.username == current_username):
             db.session.delete(model)
             db.session.commit()
             return jsonify(modelId=modelId)
